move hashidentifiedelement into planning
